+++ Com.Drew/Com/Drew/Metadata/Xmp/XmpDescriptor.cs
@@ -187,14 +187,14 @@ namespace Com.Drew.Metadata.Xmp
 			// thanks also to Gli Blr for spotting this bug
 			if (value <= 1)
 			{
-				float apexPower = (float)(1 / (Math.Exp(value * Math.Log(2))));
+				float apexPower = (float)(1 / (Math.Exp((double)value * Math.Log(2))));
 				long apexPower10 = (long)System.Math.Round((double)apexPower * 10.0);
 				float fApexPower = (float)apexPower10 / 10.0f;
 				return fApexPower + " sec";
 			}
 			else
 			{
-				int apexPower = (int)((Math.Exp(value * Math.Log(2))));
+				int apexPower = (int)((Math.Exp((double)value * Math.Log(2))));
 				return "1/" + apexPower + " sec";
 			}
 		}
